Question_14_Twitter/requirements.txt
tweepy==4.14.0
pandas
python-dotenv
configparser
pyodbc
textblob
matplotlib
seaborn